requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.41.0
deprecation==2.1.0
dj-database-url==2.2.0
Django==5.0.7
django-cloudinary-storage==0.3.0
gotrue==2.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
packaging==24.1
pillow==10.4.0
postgrest==0.16.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
django-tinymce==4.1.0
django-ckeditor



